Adding quiet flag to service account command.
Before adding this flag, the project creator would have to monitor the cloud shell to say Y to grant permission to add the service account to the project. The quiet command sets it to the default which is currently Yes. A change would be required if the default changed to No.